Handle ^D correctly and elegantly
Do not use a separate boolean flag, but rather the closedness of the child's
stdin pipe. After all, if the shell's input pipe is closed, we cannot expect
more keyboard input anyways.

More importantly, this also solves the bug of ^D handling. Previously this did
not work because the original C program rely on closing the pipe to notify the
shell to exit.