Correct handling of broken pipe
When we ask the shell to run `sleep 10; exit` during which we produce a huge
amount of input that exceeds the pipe capacity, we will be blocked in a
write(2) call to the shell. However, the shell does not read from the pipe; it
exits after ten seconds. We will get back an EPIPE + SIGPIPE + SIGCHLD.

This revision handles EPIPE correctly by closing the pipe and breaking out of
the loop to print the shell exit status. Both SIGPIPE and SIGCHLD are
effectively ignored.

The rust version is modified accordingly, except that SIGPIPE is by default
ignored.