Huge rewrite to use non-blocking fd
As noted on my Twitter, previously I didn't intuitively understand the value
of O_NONBLOCK. To me it seems like with the use of select/poll or their
replacements like epoll/kqueue we can already determine whether we can
read/write a file descriptor without blocking. But in fact I realized that the
convoluted control flow and the litany of poorly-thought-out boolean flags can
actually be replaced if we just used O_NONBLOCK everywhere. The point is *not*
keeping tracking the readiness of the file descriptors yourselves with
complicated logic, but to offload those tasks to the kernel. When you want to
write, just write till EWOULDBLOCK; when you want to read, just read till
EWOULDBLOCK.